pyproject.toml
[tool.poetry]
name = "book-agent"
version = "0.1.0"
description = ""
authors = ["gmarcinek <grzegorz.marcinek@gmail.com>"]
packages = [
    { include = "agent" },
    { include = "orchestrator" },
    { include = "ner" },
    { include = "llm" },
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
openai = ">=1.79.0,<2.0.0"
pydantic = ">=2.11.4,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
anthropic = "^0.52.2"
spacy = "^3.8.7"
transformers = "^4.52.4"
torch = "^2.7.1"
protobuf = "^6.31.1"
psutil = "^7.0.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0" 
striprtf = "^0.0.26"

[tool.poetry.scripts]
app = "orchestrator.main:main"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"